arkive-backend/translation-service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langdetect
fastkafka
aiokafka
pydantic>=2.0.0
httpx